[fix] Fix rule name mismatches between filenames and rulenames.
Also adds Apache license headers to all the AIP-203 rules,
which somehow were missing.

Prerequisite for merging #318.
