+++ rules/aip0203/optional.go
@@ -1,3 +1,17 @@
+// Copyright 2019 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package aip0203
 
 import (
@@ -14,44 +28,6 @@ var optional = &lint.FieldRule{
 	OnlyIf: withoutOptionalFieldBehavior,
 	LintField: func(f *desc.FieldDescriptor) []lint.Problem {
 		return checkLeadingComments(f, optionalRegexp, "OPTIONAL")
-	},
-}
-
-var optionalBehaviorConflict = &lint.FieldRule{
-	Name: lint.NewRuleName(203, "optional-conflict"),
-	OnlyIf: func(f *desc.FieldDescriptor) bool {
-		return !withoutOptionalFieldBehavior(f)
-	},
-	LintField: func(f *desc.FieldDescriptor) []lint.Problem {
-		// APIs may use the OPTIONAL value to describe a field which doesn't use
-		// REQUIRED, IMMUTABLE, INPUT_ONLY or OUTPUT_ONLY. If a field is described
-		// as optional, it can't be others.
-		if len(utils.GetFieldBehavior(f)) > 1 {
-			return []lint.Problem{{
-				Message:    "Field behavior `(google.api.field_behavior) = OPTIONAL` shouldn't be used together with other field behaviors.",
-				Descriptor: f,
-			}}
-		}
-		return nil
-	},
-}
-
-// If a message has a field which is described as optional, ensure that every
-// optional field on the message has this indicator.
-var optionalBehaviorConsistency = &lint.MessageRule{
-	Name:   lint.NewRuleName(203, "optional-consistency"),
-	OnlyIf: messageHasOptionalFieldBehavior,
-	LintMessage: func(m *desc.MessageDescriptor) (problems []lint.Problem) {
-
-		for _, f := range m.GetFields() {
-			if utils.GetFieldBehavior(f) == nil {
-				problems = append(problems, lint.Problem{
-					Message:    "Within a single message, either all optional fields should be indicated, or none of them should be.",
-					Descriptor: f,
-				})
-			}
-		}
-		return
 	},
 }
